Skip deleting files that we just deleted
We see this happening with Swift. Because the consistency of swift is eventual, swift sometimes didn't process the deletion of the meta file yet, and so it turns up in the bkt.Iter(). The second deletion then causes a 404 and compaction fails.

+++ pkg/block/block.go
@@ -156,6 +156,11 @@ func deleteDir(ctx context.Context, logger log.Logger, bkt objstore.Bucket, dir 
 		if strings.HasSuffix(name, objstore.DirDelim) {
 			return deleteDir(ctx, logger, bkt, name)
 		}
+		metaFile := path.Join(id.String(), MetaFilename)
+		if name == metaFile {
+			// skip this file if we found it, it was already deleted
+			continue
+		}
 		if err := bkt.Delete(ctx, name); err != nil {
 			return err
 		}
